requirements.txt
Flask
matplotlib
fpdf
python-dotenv
pymongo
google-generativeai
gunicorn